Ego4d/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.3.0=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.3.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
av=12.0.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
black=21.4b2=pypi_0
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
decord=0.6.0=pypi_0
defusedxml=0.7.1=pypi_0
detectron2=0.6=pypi_0
deva=1.0.0=pypi_0
distro=1.9.0=pypi_0
editdistance=0.6.0=pypi_0
einops=0.3.0=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fastapi=0.110.0=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.13.3=pypi_0
fonttools=4.50.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
future=1.0.0=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gdown=5.1.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gradio=4.24.0=pypi_0
gradio-client=0.14.0=pypi_0
grpcio=1.62.1=pypi_0
gurobipy=11.0.1=pypi_0
h11=0.14.0=pypi_0
h5py=3.10.0=pypi_0
hickle=5.0.3=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.22.2=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.6=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
imutils=0.5.4=pypi_0
iopath=0.1.9=pypi_0
ipdb=0.13.13=pypi_0
ipython=8.18.1=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lightning-utilities=0.11.2=pypi_0
lmdb=1.2.1=pypi_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.99=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
omegaconf=2.3.0=pypi_0
openai=1.14.3=pypi_0
opencv-python=4.5.3.56=pypi_0
opencv-python-headless=4.9.0.80=pypi_0
openssl=3.0.13=h7f8727e_0
orjson=3.10.0=pypi_0
packaging=24.0=pypi_0
pandas=1.3.5=pypi_0
parameterized=0.9.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.12.1=pypi_0
pexpect=4.9.0=pypi_0
pillow=9.5.0=pypi_0
pip=23.3.1=py39h06a4308_0
portalocker=2.8.2=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=5.26.1=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pypi_0
pulp=2.8.0=pypi_0
pure-eval=0.2.2=pypi_0
pycocotools=2.0.7=pypi_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pydeprecate=0.3.1=pypi_0
pydot=2.0.0=pypi_0
pydub=0.25.1=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=pypi_0
python=3.9.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.9=pypi_0
pytorch-lightning=1.5.6=pypi_0
pytorchvideo=0.1.5=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.34.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.0=pypi_0
ruff=0.3.5=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.12.0=pypi_0
segment-anything=1.0=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=59.5.0=pypi_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
starlette=0.36.3=pypi_0
submitit=1.3.3=pypi_0
supervision=0.19.0=pypi_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.4.0=pypi_0
thinplate=1.0.0=pypi_0
threadpoolctl=3.4.0=pypi_0
timm=0.9.16=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2=pypi_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
torch=1.9.0+cu111=pypi_0
torchaudio=0.9.0=pypi_0
torchmetrics=1.2.1=pypi_0
torchvision=0.10.0+cu111=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.2=pypi_0
transformers=4.39.3=pypi_0
triton=2.2.0=pypi_0
typer=0.12.0=pypi_0
typer-cli=0.12.0=pypi_0
typer-slim=0.12.0=pypi_0
typing-extensions=4.10.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.29.0=pypi_0
wcwidth=0.2.13=pypi_0
websockets=11.0.3=pypi_0
werkzeug=3.0.2=pypi_0
wheel=0.41.2=py39h06a4308_0
xz=5.4.6=h5eee18b_0
yacs=0.1.8=pypi_0
yarl=1.9.4=pypi_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_0
